Refactor: split sql reader and sql writer tests.
