+++ pkg/pgmodel/ingestor/buffer.go
@@ -1,4 +1,4 @@
-package ingester
+package ingestor
 
 import (
 	"sync"
