+++ pkg/pgmodel/ingestor/ingestor.go
@@ -1,4 +1,4 @@
-package ingester
+package ingestor
 
 import (
 	"fmt"
