+++ pkg/pgmodel/ingestor/ingestor_test.go
@@ -1,4 +1,4 @@
-package ingester
+package ingestor
 
 import (
 	"fmt"
