+++ pkg/pgmodel/ingestor/insert_ctx.go
@@ -1,4 +1,4 @@
-package ingester
+package ingestor
 
 import (
 	"sync"
