+++ pkg/tests/end_to_end_tests/concurrent_sql_test.go
@@ -6,7 +6,7 @@ package end_to_end_tests
 import (
 	"context"
 	"fmt"
-	"github.com/timescale/promscale/pkg/pgmodel/ingester"
+	"github.com/timescale/promscale/pkg/pgmodel/ingestor"
 	"github.com/timescale/promscale/pkg/pgmodel/utils"
 	"strings"
 	"sync"
@@ -221,12 +221,12 @@ func testConcurrentInsertSimple(t testing.TB, db *pgxpool.Pool, metric string) {
 		},
 	}
 
-	ingestor, err := ingester.NewPgxIngestor(pgxconn.NewPgxConn(db))
+	ingestor, err := ingestor.NewPgxIngestor(pgxconn.NewPgxConn(db))
 	if err != nil {
 		t.Fatal(err)
 	}
 	defer ingestor.Close()
-	_, err = ingestor.Ingest(copyMetrics(metrics), ingester.NewWriteRequest())
+	_, err = ingestor.Ingest(copyMetrics(metrics), ingestor.NewWriteRequest())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -280,13 +280,13 @@ func testConcurrentInsertAdvanced(t testing.TB, db *pgxpool.Pool) {
 		},
 	}
 
-	ingestor, err := ingester.NewPgxIngestor(pgxconn.NewPgxConn(db))
+	ingestor, err := ingestor.NewPgxIngestor(pgxconn.NewPgxConn(db))
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	defer ingestor.Close()
-	_, err = ingestor.Ingest(copyMetrics(metrics), ingester.NewWriteRequest())
+	_, err = ingestor.Ingest(copyMetrics(metrics), ingestor.NewWriteRequest())
 	if err != nil {
 		t.Fatal(err)
 	}
